# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  version: 1.0.0
  description: Connect with your friends effortlessly using WASAText!
servers:
  - url: http://localhost:8080/v1
...
paths:
...
  /conversations/{conversationId}/messages:
    post:
      operationId: sendMessage
      summary: Send message
      description: Sends a message to a conversation
      tags:
        - conversations
      parameters:
        - $ref: "#/components/parameters/conversationId"
      requestBody:
        description: Message details
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Message object
              properties:
                content:
                  $ref: "#/components/schemas/Content"
                image:
                  $ref: "#/components/schemas/Image"
              required:
                - content
            example:
              content: Hello, how are you?
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              examples:
                messageExample:
                  $ref: "#/components/examples/messageExample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /conversations/{conversationId}/forwards:
    post:
      operationId: forwardMessage
      summary: Forward message
      description: Forwards a message to a conversation
      tags:
        - conversations
      parameters:
        - $ref: "#/components/parameters/conversationId"
      requestBody:
        description: Message details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Forwarded message object
              properties:
                messageId:
                  $ref: "#/components/schemas/Id"
              required:
                - messageId
            example:
              messageId: "550e8400-e29b-41d4-a716-446655440000"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              examples:
                messageExample:
                  $ref: "#/components/examples/messageExample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/messageId"
    put:
      operationId: editMessage
      summary: Update message
      description: Update the content of a message
      tags:
        - conversations
      requestBody:
        description: Message details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Content object
              properties:
                content:
                  $ref: "#/components/schemas/Content"
              required:
                - content
            example:
              content: Hello, how are you?
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Message updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              examples:
                messageExample:
                  $ref: "#/components/examples/messageExample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteMessage
      summary: Delete message
      description: Permanently deletes a message
      tags:
        - conversations
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Message deleted successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /messages/{messageId}/comments:
    post:
      operationId: commentMessage
      summary: Comment message
      description: Adds a comment to a message
      tags:
        - conversations
      parameters:
        - $ref: "#/components/parameters/messageId"
      requestBody:
        description: Comment details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Emoji object
              properties:
                emoji:
                  $ref: "#/components/schemas/Emoji"
              required:
                - emoji
            example:
              emoji: "😀"
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
              examples:
                commentExample:
                  $ref: "#/components/examples/commentExample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /comments/{commentId}:
    delete:
      operationId: uncommentMessage
      summary: Uncomment message
      description: Permanently deletes a comment from a message
      tags:
        - conversations
      parameters:
        - $ref: "#/components/parameters/commentId"
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Comment deleted successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  schemas:
...
    Id:
      type: string
      format: uuid
      minLength: 36
      maxLength: 36
      pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
      description: Unique identifier
    Timestamp:
      type: string
      format: date-time
      minLength: 20
      maxLength: 20
      description: Timestamp in RFC 3339 format
    Image:
      type: string
      format: binary
      minLength: 1
      maxLength: 4194304
      description: Binary data for image upload

    # --------------------------------------------------------------------------------
    # User
    Username:
      type: string
      minLength: 3
      maxLength: 16
      pattern: "^[a-zA-Z0-9._]{3,16}$"
      description: Unique handle used to identify the user
    ProfilePicture:
      type: string
      minLength: 11
      maxLength: 255
      pattern: "^http://(localhost|127\\.0\\.0\\.1):[0-9]{1,5}/uploads/profile-pictures/(default|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})\\.(jpg|jpeg|png|gif|webp|bmp|svg)$"
      description: URL of the user's profile picture
    UserPreview:
      type: object
      description: Preview of user details
      properties:
        userId:
          $ref: "#/components/schemas/Id"
        username:
          $ref: "#/components/schemas/Username"
        profilePicture:
          $ref: "#/components/schemas/ProfilePicture"
      required:
        - userId
        - username
...
    Content:
      type: string
      minLength: 1
      maxLength: 1000
      pattern: "^.*$"
      description: Message content
    Attachment:
      type: string
      minLength: 11
      maxLength: 255
      pattern: "^http://(localhost|127\\.0\\.0\\.1):[0-9]{1,5}/uploads/attachments/[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}\\.(jpg|jpeg|png|gif|webp|bmp|svg)$"
      description: URL of the attachment
    Message:
      type: object
      description: Message details
      properties:
        messageId:
          $ref: "#/components/schemas/Id"
        content:
          $ref: "#/components/schemas/Content"
        attachment:
          $ref: "#/components/schemas/Attachment"
        isForwarded:
          type: boolean
          description: Indicates if the message is forwarded
        sender:
          $ref: "#/components/schemas/UserPreview"
        status:
          type: object
          description: Message status
          properties:
            delivered:
              type: object
              description: Delivery status of the message
              additionalProperties:
                $ref: "#/components/schemas/Timestamp"
            read:
              type: object
              description: Read status of the message
              additionalProperties:
                $ref: "#/components/schemas/Timestamp"
          required:
            - delivered
            - read
        comments:
          type: array
          minItems: 0
          maxItems: 100
          description: List of comments on the message
          items:
            $ref: "#/components/schemas/Comment"
        sentAt:
          $ref: "#/components/schemas/Timestamp"
        editedAt:
          $ref: "#/components/schemas/Timestamp"
      required:
        - messageId
        - content
        - isForwarded
        - sender
        - status
        - comments
        - sentAt

    # --------------------------------------------------------------------------------
    # Comment
    Emoji:
      type: string
      minLength: 1
      maxLength: 10
      pattern: "^[\u1F600-\u1F64F\u1F300-\u1F5FF\u1F680-\u1F6FF\u2600-\u26FF\u2700-\u27BF]+$"
      description: Emoji used in the comment
    Comment:
      type: object
      description: Comment details
      properties:
        commentId:
          $ref: "#/components/schemas/Id"
        commenter:
          $ref: "#/components/schemas/UserPreview"
        emoji:
          $ref: "#/components/schemas/Emoji"
        commentedAt:
          $ref: "#/components/schemas/Timestamp"
      required:
        - commentId
        - emoji
        - commentedAt

    # --------------------------------------------------------------------------------
    # Utils
    Error:
      type: object
      description: Error details
      properties:
        code:
          type: string
          minLength: 7
          maxLength: 7
          pattern: "^ERR-[0-9]{3}$"
          description: Machine-readable error code
        message:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^.*$"
          description: Brief description of the error
      required:
        - code
        - message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "ERR-400"
            message: Bad request
    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "ERR-401"
            message: Unauthorized access
    Forbidden:
      description: Forbidden access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "ERR-403"
            message: Forbidden access
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "ERR-404"
            message: Resource not found
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "ERR-500"
            message: Internal server error
  examples:
...
    messageExample:
      summary: Example message
      value:
        messageId: "550e8400-e29b-41d4-a716-446655440000"
        content: "Hello, how are you?"
        isForwarded: false
        sender:
          userId: "550e8400-e29b-41d4-a716-446655440000"
          username: Maria
          profilePicture: "http://localhost:8080/uploads/profile-pictures/550e8400-e29b-41d4-a716-446655440000.jpg"
        status:
          delivered:
            "550e8400-e29b-41d4-a716-446655440000": "2023-10-01T12:00:00Z"
          read:
            "550e8400-e29b-41d4-a716-446655440000": "2023-10-01T12:05:00Z"
        comments: []
        sentAt: "2023-10-01T12:00:00Z"
    commentExample:
      summary: Example comment
      value:
        commentId: "550e8400-e29b-41d4-a716-446655440000"
        emoji: "😀"
        commentedAt: "2023-10-01T12:10:00Z"
  parameters:
...
    conversationId:
      name: conversationId
      in: path
      required: true
      description: Unique identifier of the conversation
      schema:
        $ref: "#/components/schemas/Id"
      example: "550e8400-e29b-41d4-a716-446655440000"
    messageId:
      name: messageId
      in: path
      required: true
      description: Unique identifier of the message
      schema:
        $ref: "#/components/schemas/Id"
      example: "550e8400-e29b-41d4-a716-446655440000"
    commentId:
      name: commentId
      in: path
      required: true
      description: Unique identifier of the comment
      schema:
        $ref: "#/components/schemas/Id"
      example: "550e8400-e29b-41d4-a716-446655440000"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: uuid